Update README.md
The laptop market is a competitive one, with many different brands and models to choose from. This can make it difficult for consumers to make informed decisions about which laptop to buy. The dataset I cleaned and analyzed can help consumers to make better decisions by providing them with information about the prices, specifications, and features of different laptops.